Mourad_NAJI_IAODC_201_CC3_Projet_NLP/requirements.txt
fastapi==0.134.0
uvicorn==0.41.0
pydantic==2.12.5
pydantic_core==2.41.5
starlette==0.52.1
python-dotenv
requests==2.32.5
httpx
sentence-transformers==2.7.0
transformers==4.40.2
faiss-cpu==1.8.0
torch==2.2.2
numpy==1.26.4
pandas==2.3.3
pillow==12.1.1
tenacity==9.1.4
typing_extensions==4.15.0

